requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
azure-core==1.30.2
azure-identity==1.17.1
azure-storage-blob==12.22.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Deprecated==1.2.14
Flask==3.0.3
gunicorn==23.0.0
idna==3.8
importlib_metadata==8.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.31.0
msal-extensions==1.2.0
opentelemetry-api==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-wsgi==0.48b0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
packaging==24.1
portalocker==2.10.1
pycparser==2.22
pycryptodomex==3.20.0
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.1
